+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Add
-- (pro) Add encrypted source URLs support.
+- (pro) Add encrypted source URL support.
 
 ## [3.7.2] - 2022-08-22
 ### Changed